switch DTO id to int for auto incrementing

+++ uTransnet-Calc/Assets/uTrans/Scripts/Data/BaseDTO.cs
@@ -1,7 +1,9 @@
+using System;
+
 namespace uTrans.Data
 {
     public interface BaseDTO
     {
-        long Id { get; set; }
+        int Id { get; set; }
     }
 }
