+++ uTransnet-Calc/Assets/uTrans/Scripts/Data/PresetMaterialDTO.cs
@@ -5,9 +5,9 @@ namespace uTrans.Data
     public class PresetMaterialDTO : BaseDTO
     {
         [PrimaryKey, AutoIncrement]
-        public long Id { get; set; }
-        public long PresetId { get; set; }
-        public long MaterialId { get; set; }
+        public int Id { get; set; }
+        public int PresetId { get; set; }
+        public int MaterialId { get; set; }
         public int Price { get; set; }
     }
 }
